+++ src/app/features/home/capture-tab/capture-details/capture-details.page.ts
@@ -62,7 +62,8 @@ export class CaptureDetailsPage {
       const [index, meta] = Object.entries(proof.indexedAssets)[0];
       if (!(await this.imageStore.exists(index)) && proof.diaBackendAssetId) {
         const imageBlob = await this.diaBackendAssetRepository
-          .downloadFile$(proof.diaBackendAssetId, 'asset_file')
+          .downloadFile$({ id: proof.diaBackendAssetId, field: 'asset_file' })
+          .pipe(first())
           .toPromise();
         await proof.setAssets({ [await blobToBase64(imageBlob)]: meta });
       }
@@ -110,14 +111,13 @@ export class CaptureDetailsPage {
       data: { email: '' },
     });
     const contact$ = dialogRef.afterClosed().pipe(isNonNullable());
-    combineLatest([contact$, this.proof$])
+
+    return combineLatest([contact$, this.proof$])
       .pipe(first(), untilDestroyed(this))
       .subscribe(([contact, proof]) =>
         this.router.navigate(
           ['sending-post-capture', { contact, id: proof.diaBackendAssetId }],
-          {
-            relativeTo: this.route,
-          }
+          { relativeTo: this.route }
         )
       );
   }
@@ -209,7 +209,7 @@ export class CaptureDetailsPage {
       switchTap(proof =>
         defer(() => {
           if (proof.diaBackendAssetId) {
-            return this.diaBackendAssetRepository.removeById$(
+            return this.diaBackendAssetRepository.removeCaptureById$(
               proof.diaBackendAssetId
             );
           }
@@ -229,7 +229,7 @@ export class CaptureDetailsPage {
   }
 }
 
-function isValidGeolocation(proof: Proof) {
+export function isValidGeolocation(proof: Proof) {
   return proof.geolocationLatitude === undefined ||
     proof.geolocationLatitude === 'undefined' ||
     proof.geolocationLongitude === undefined ||