+++ src/app/features/home/home.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { PostCaptureCardModule } from '../../shared/core/post-capture-card/post-capture-card.module';
 import { SharedModule } from '../../shared/shared.module';
 import { CaptureTabComponent } from './capture-tab/capture-tab.component';
 import { UploadingBarComponent } from './capture-tab/uploading-bar/uploading-bar.component';
@@ -9,7 +8,7 @@ import { PrefetchingDialogComponent } from './onboarding/prefetching-dialog/pref
 import { PostCaptureTabComponent } from './post-capture-tab/post-capture-tab.component';
 
 @NgModule({
-  imports: [SharedModule, HomePageRoutingModule, PostCaptureCardModule],
+  imports: [SharedModule, HomePageRoutingModule],
   declarations: [
     HomePage,
     CaptureTabComponent,